Move fuzzing tests into their own file. apply some comments from review
